(refactor) Sending SMS is extracted into a service object
* This will allow us to blacklist NBT phone numbers from receiving text messages in a single place.
